+++ package.json
@@ -8,8 +8,7 @@
   },
   "scripts": {
     "start": "node --experimental-strip-types index.ts",
-    "build": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc"
   },
   "keywords": [],
   "author": "",
